Add a function to provide details of content through content id

+++ src/main/kotlin/com/aroundme/content/ContentRepository.kt
@@ -2,6 +2,6 @@ package com.aroundme.content
 
 import org.springframework.data.repository.CrudRepository
 
-interface ContentRepository: CrudRepository<Content, Int> {
+interface ContentRepository: CrudRepository<Content, Long> {
     fun findAllBy(): List<Content>
 }
